Updated redirect to check if a redirect is neccessary
